fix: Время работы/останова

+++ src/user/device/ccs/ccs.cpp
@@ -574,8 +574,8 @@ void Ccs::calcTime()
     conditionOld = condition;
   }
 
-  setNewValue(CCS_RUN_TIME, getSecFromCurTime(CCS_LAST_RUN_DATE_TIME));
-  setNewValue(CCS_STOP_TIME, getSecFromCurTime(CCS_LAST_STOP_DATE_TIME));
+  setNewValue(CCS_RUN_TIME, (float)getSecFromCurTime(CCS_LAST_RUN_DATE_TIME));
+  setNewValue(CCS_STOP_TIME, (float)getSecFromCurTime(CCS_LAST_STOP_DATE_TIME));
 }
 
 void Ccs::calcParameters()
